ProfilePopUp completed, useClickedOutsideEffect hook added
